style/PEP8-style

+++ backend/explorer/calculators.py
@@ -1,6 +1,6 @@
 from post.models import Post
 
 
-def top_post_formula(post: Post) -> int:
+def top_post_formula(post: Post) -> float:
     result = (post.likes / post.views) * 100
     return result
